requirements.txt
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
et-xmlfile==1.1.0
google-api-core==2.11.1
google-api-python-client==2.93.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
httplib2==0.22.0
idna==3.4
mysql-connector-python==8.1.0
numpy==1.25.1
oauth2client==4.1.3
openpyxl==3.1.2
pandas==2.0.3
pdfminer.six==20221105
protobuf==4.21.12
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyDrive==1.3.1
pyparsing==3.1.0
PyPDF2==1.26.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.31.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
